pyproject.toml
[project]
name = "neura-vault-backend"
version = "0.1.0"
description = "A robust Django-based backend service for DeFi yield optimization and allocation"
readme = "README.md"
authors = [
    {name = "Nura Vault Team"}
]
requires-python = ">=3.11"
dependencies = [
    "black>=25.1.0",
    "boto3>=1.28.0",
    "crewai>=0.120.1",
    "crewai-tools>=0.45.0",
    "django>=5.2.1",
    "django-cors-headers>=4.7.0",
    "django-environ>=0.10.0",
    "django-filter>=24.1",
    "djangorestframework>=3.16.0",
    "dj-database-url>=2.1.0",
    "drf-spectacular>=0.28.0",
    "gevent>=24.2.1",
    "gunicorn>=21.2.0",
    "jq>=1.8.0",
    "langchain-anthropic>=0.3.13",
    "langchain-chroma>=0.2.3",
    "psutil>=5.9.8",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "scrapy>=2.13.0",
    "web3>=6.15.1",
    "litellm>=1.15.0",
    "openai>=1.13.3",
    "httpx>=0.25.1",
    "asyncio>=3.4.3",
]
